Add student command
The Student is a Person with StudentID and Tutorial attributes.
For now, we assume that a student only belongs in 1 tutorial.
The way a Student is saved is different from a Person
(JSON stores more fields for id and tutorial).

For listing, a Student may also be displayed different from a Person.